+++ NEWS.md
@@ -7,6 +7,7 @@
 - fix vertical overlapping lines with grid output (issue #644)
 - fix broken internal links in PDF file, probably due to a change in knitr or 
 rmarkdown (issue #632)
+- fix right outer border issue in grid format (issue #650)
 
 # flextable 0.9.6